MoviePage UI improvements and Navbar refactor
